requirements.txt
pyserial
matplotlib
scipy
seaborn
scikit-learn
scipy
pywavelets